Update the Testnet rewards requirements
- Add longevity description
- Add more explicit active bid requirement
- Update the network weight threshold
- Update the uptime block buffer to 4

These have been well-documented in various places,
and communicated through the official comms channel
on Telegram before. Although we haven't got any complaints
about anyone not knowing these rules, based on community feedback,
we are adding here explicitly to make it more accessible
and easier to notice for the newcomers.

Signed-off-by: Karol Swiderski <karol.swiderski.dev@gmail.com>